`tmpnet`: Reuse dynamically-allocated API port across restarts
Prometheus labels collected metrics with an `instance`
value (host:port) and if the port changes across restarts then the
collected metrics for a node end up segmented by the different
instance value. Reusing an ephemeral port is not 100% safe but the
alternative option of special-casing all the grafana dashboard queries
to aggregate across instance values is arguably worse.

This isn't a problem for avalanchego testing today, but testing that
involve subnets and the node restarts that subnet setup
entails (e.g. xsvm, subnet-evm, etc) need this fix to allow grafana to
display results unambiguously.